backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
cachetools==6.2.2
    # via google-auth
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
    #   uvicorn
exceptiongroup==1.3.1
    # via
    #   anyio
    #   pytest
fastapi==0.122.0
    # via -r requirements.in
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.28.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.187.0
    # via google-generativeai
google-auth==2.43.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.1
    # via google-api-python-client
google-generativeai==0.8.5
    # via -r requirements.in
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.3.0
    # via pytest
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via pytest
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.4
    # via
    #   fastapi
    #   google-generativeai
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via pytest
pyparsing==3.2.5
    # via httplib2
pytest==9.0.1
    # via -r requirements.in
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.3
    # via uvicorn
requests==2.32.5
    # via google-api-core
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
starlette==0.50.0
    # via fastapi
tomli==2.3.0
    # via pytest
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   google-generativeai
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via -r requirements.in
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
